Fix bug in RegSetString
There was a bug that would only ever write the first 13 characters in a
string to the registry.
The problem was the use of unsafe.Sizeof(buf) with buf being a []byte.
This always returns the size of a slice header, I assume. Actually it
should have been the length of the buffer all along.
Also the null-terminator is appended by syscall.UTF16FromString
automatically, so there is no need for the +2 in the code.