Add functions to draw polygons from memory
Instead of using []POINT slices there are now versions that accept
memory locations. This allows optimizations for when data is in the
right form (two 32 bit integers) but has the wrong static type, e.g. a
custom Point struct instead of w32.POINT. Using the memory pointer
functions we do not need to copy the array just for the right type, we
can pass it as is.