Restore v1.0.0 files in root folder
This repository was not created with Go modules or semantic versioning
in mind. Now that Go 1.16 uses modules by default, we try to re-create
what would have been here if we had used modules and sematic versioning
from the start.

The latest stuff is now in v2 with tags starting at v2.3.0. The root
folder was reverted to the v1.0.0 code.

This messes up the git version history, it will not look right and
things might be hard to find. However going forward we will pay
attention to what changes are breaking the existing API and create a
v3, v4, etc. in these cases. Hopefully this will make this repo future
proof, we sacrafice a little convenience right now for it.